Worked on training data collection
Now (hopefully) creates a proper training dataset.
Added a variable to choose whether to recreate/update the
training data.
Program now fakes being a browser in the header for url requests
(This was causing issues because a bunch of sites respond with
an error otherwise.  Interestingly it was mostly the sites in
the trainingBad).
Training html stored as .txt in a specific folder structure
to utilize sklearn.datasets.load_files for making the dataset.