Moved some stuff to htmlToTxtMaker.py Added htmlClassifier.py
We essentially now have the trainingCorpus with categories initialized.
We need to train a classifier (probably nltk naive bayes cause scikit svm is
proving to be a bust) and from their we should be able to throw tokenized
text at the classifier and determine which category it should belong to.